BuzzBuzzGPTs-STT/requirements.txt
# Production-locked dependencies for BuzzBuzzGPTs-STT
# Verified on Python 3.13.1

anyio==4.12.1
av==16.1.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
ctranslate2==4.6.3
customtkinter==5.2.2
darkdetect==0.8.0
faster-whisper==1.2.1
filelock==3.20.3
flatbuffers==25.12.19
fsspec==2026.1.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.5
humanfriendly==10.0
idna==3.11
Jinja2==3.1.6
llvmlite==0.46.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.6.1
numba==0.63.1
numpy==1.26.4
onnxruntime==1.23.2
openai-whisper==20250625
packaging==26.0
protobuf==6.33.4
pycparser==3.0
Pygments==2.19.2
pyreadline3==3.5.4
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
rich==14.3.1
setuptools==80.10.2
shellingham==1.5.4
SoundCard==0.4.5
sympy==1.14.0
tiktoken==0.12.0
tokenizers==0.22.2
torch==2.10.0
torchaudio==2.10.0
tqdm==4.67.1
typer-slim==0.21.1
typing_extensions==4.15.0
urllib3==2.6.3
